# doc/api.yml
openapi: 3.0.3
info:
  version: '1.0'
  title: HEI Admin API
  description: |
    _Programmatically connect to a computer programming [school](https://hei.school)._
    After [joining us](mailto:contact@hei.school),
    you can get an identification token from
    our [dev](https://dev-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=5s8cg50doahmu855rlc8fr6qmp&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-dev.hei.school%2Fwhoami)
    or [prod](https://prod-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=i8bg538jpfu6mqmqb61m26trd&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-prod.hei.school%2Fwhoami)
    authentication service.
    Then, start playing with our system!

    The implementation of our API is [publicly disclosed](https://github.com/hei-school/hei-admin-api).
    You are welcome to try and compromise it. Happy hacking!
servers:
  - url: 'https://api-dev.hei.school'
  - url: 'https://api-prod.hei.school'
security:
  - BearerAuth: [ ]
paths:
...
  '/groups/{group_id}/students':
    get:
      tags:
        - Users
      summary: Get students by groups
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
      operationId: getStudentsByGroupId
      responses:
        '200':
          description: The students list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
  '/students/{id}':
    get:
      tags:
        - Users
      summary: Get student by identifier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      operationId: getStudentById
      responses:
        '200':
          description: The identified student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /students:
    get:
      tags:
        - Users
      summary: Get all students
      operationId: getStudents
      parameters:
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: ref
          in: query
          description: "Filter students by ref, case is ignored"
          schema:
            type: string
        - name: first_name
          in: query
          description: "Filter students by first name, case is ignored"
          schema:
            type: string
        - name: last_name
          in: query
          description: "Filter students by last name, case is ignored"
          schema:
            type: string
      responses:
        '200':
          description: List of students, ordered by ref.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Users
      summary: Create new students or update existing students
      description: Update students when `id` are provided, create them otherwise.
      operationId: createOrUpdateStudents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: The created or updated students
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
  '/events/{id}':
    get:
      tags:
        - Event
      summary: Get event by identifier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      operationId: getEventById
      responses:
        '200':
          description: The identified event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /events:
    get:
      tags:
        - Event
      summary: Get all events order by starting_hours
      parameters:
        - name: place_id
          in: query
          required: false
          schema:
            type: string
      operationId: getEvents
      responses:
        '200':
          description: List of all events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Event
      summary: Create new events or update existing events
      description: Update events when id are provided, create them otherwise.
      operationId: createOrUpdateEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: The created or updated events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  '/events/{event_id}/event_participants/{event_participant_id}':
    get:
      tags:
        - Event
      summary: Get event event_participant by identifier
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
        - name: event_participant_id
          in: path
          required: true
          schema:
            type: string
      operationId: getEventEventParticipantById
      responses:
        '200':
          description: Identified event event participant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventParticipant'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  '/events/{event_id}/event_participants':
    get:
      tags:
        - Event
      summary: Get all event event_participants
      description: Event participants can be filtered with an optional status parameter.
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum:
              - EXPECTED
              - HERE
              - MISSING
      operationId: getEventEventParticipants
      responses:
        '200':
          description: List of all event event_participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventParticipant'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    post:
      tags:
        - Event
      summary: Create new event event_participants
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      operationId: createEventEventParticipants
      requestBody:
        description: Event event participants to create
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateEventParticipant'
      responses:
        '200':
          description: Created Event event participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventParticipant'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Event
      summary: Update existing event event_participants
      description: Update event event_participants when id are provided, create them otherwise.
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      operationId: updateEventEventParticipants
      requestBody:
        description: Event event participants to update
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/EventParticipant'
      responses:
        '200':
          description: Updated Event event participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventParticipant'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /events/{event_id}/presence:
    put:
      tags:
        - Event
      summary: Update a event_participant's status with their image.
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      operationId: updateEventEventParticipantsStatus
      requestBody:
        description: Event event participants to update
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Updated Event event participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventParticipant'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /event_participants:
    get:
      tags:
        - Event
      summary: Get all event event_participants filtered by status
      parameters:
        - name: status
          in: query
          description: "See the event_participants object for its value."
          schema:
            type: string
            default: EXPECTED
            enum:
              - EXPECTED
              - HERE
              - MISSING
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
      operationId: getEventParticipants
      responses:
        '200':
          description: List of filtered event event_participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventParticipant'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    User:
      properties:
        id:
          type: string
        ref:
          type: string
          example: STD000001
        first_name:
          type: string
        last_name:
          type: string
        sex:
          type: string
          enum:
            - M
            - F
        birth_date:
          type: string
          format: date
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        entrance_datetime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/EnableStatus'
        key_image_in_bucket:
          type: string
          example: toky1.jpg
    EnableStatus:
      type: string
      enum:
        - ENABLED
        - DISABLED
    Student:
      properties:
        group_id:
          type: string
      allOf:
        - $ref: '#/components/schemas/User'
...
    CreateEventParticipant:
      type: object
      properties:
        user_participant_id:
          type: string
    EventParticipant:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - EXPECTED
            - HERE
            - MISSING
        event_id:
          type: string
      allOf:
        - $ref: '#/components/schemas/CreateEventParticipant'
    Event:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: The event number one
        ref:
          type: string
          example: EVT1
        starting_hours:
          type: string
          format: date-time
        ending_hours:
          type: string
          format: date-time
        user_manager_id:
          type: string
        place_id:
          type: string
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error